+++ sample/src/main/java/ua/pp/ihorzak/alog/sample/ALogSampleApplication.java
@@ -20,6 +20,7 @@ import android.app.Application;
 
 import ua.pp.ihorzak.alog.ALog;
 import ua.pp.ihorzak.alog.ALogConfiguration;
+import ua.pp.ihorzak.alog.ALogFileConfiguration;
 
 /**
  * ALog sample application.
@@ -28,12 +29,17 @@ import ua.pp.ihorzak.alog.ALogConfiguration;
  */
 @SuppressWarnings({"WeakerAccess", "RedundantSuppression"})
 public class ALogSampleApplication extends Application {
+    private static final long LOG_FILE_LIFETIME = 3600L;
+
     @Override
     public void onCreate() {
         super.onCreate();
 
         // Initialize ALog
+        ALogFileManager aLogFileManager = new ALogFileManager(this, LOG_FILE_LIFETIME);
+        aLogFileManager.deleteObsoleteFiles();
         ALogConfiguration aLogConfiguration = ALogConfiguration.builder()
+                .file(ALogFileConfiguration.single(aLogFileManager.createFile().getPath(), false))
                 .tag("ALogSampleApplication")
                 .classPrefixEnabled(true)
                 .jsonIndentSpaceCount(4)